Migrate commands to new module, remove old bot
